auth_backend/requirements.txt
fastapi[standard]
uvicorn
PyJWT
bcrypt
sqlalchemy
psycopg2